[I8042PRT] Move DMI/SMBIOS parsing functions into library
