+++ drivers/input/i8042prt/CMakeLists.txt
@@ -1,4 +1,7 @@
 
+include_directories(
+    ${REACTOS_SOURCE_DIR}/sdk/lib/dmilib)
+
 list(APPEND SOURCE
     createclose.c
     hwhacks.c
@@ -18,6 +21,7 @@ add_library(i8042prt SHARED
     i8042prt.rc)
 
 set_module_type(i8042prt kernelmodedriver)
+target_link_libraries(i8042prt dmilib)
 add_importlibs(i8042prt ntoskrnl hal)
 add_pch(i8042prt i8042prt.h SOURCE)
 add_cd_file(TARGET i8042prt DESTINATION reactos/system32/drivers NO_CAB FOR all)
