+++ sdk/lib/CMakeLists.txt
@@ -18,6 +18,7 @@ if(MSVC)
 endif()
 
 add_subdirectory(delayimp)
+add_subdirectory(dmilib)
 #add_subdirectory(dnslib) Nothing links to this lib.
 add_subdirectory(drivers)
 add_subdirectory(dxguid)
