+++ sdk/lib/dmilib/precomp.h
@@ -1,0 +1,9 @@
+#ifndef _DMILIB_PCH_
+#define _DMILIB_PCH_
+
+#include <ntdef.h>
+#include <wmidata.h>
+#include "dmi.h"
+#include "dmilib.h"
+
+#endif /* _DMILIB_PCH_ */
